# src/infrastructure/Bathhouse.Contracts/Models/v1/ClientResponse.cs
using Bathhouse.ValueTypes;
using System;

namespace Bathhouse.Contracts.v1.Models
{
  /// <summary>
  /// 
  /// </summary>
  public record ClientResponse(
    Guid Id, 
    string LastName, 
    string MiddleName, 
    string FirstName, 
    string Phone, 
    DateTime? DoB, 
    string? Comment,
    Sex Sex,
    string OfficeNumber,
    Guid? OfficeId);
}